Support URL params and cross-API resources.
Added a property that determines whether the input is expected as JSON
in the body of the request or URL-encoded in the URL of the request.
Also set the parent definition to include the API ID when referencing
the resource, to allow cross-API resource reuse (e.g., so projects don't
need to be redefined in every single API).