Updated the MQ resources to new resource file spec.
The Queue and Message resources now match the current API (more or less)
in features, but the resource file spec needs to be expanded to match
the changes that were necessary to cover the functionality.

We also need to create a webhook resource to post messages via webhook,
a reservations resource to reserve messages from the queue and
touch/release them, and a subscriptions endpoint to manage push queues.
Though the subscriptions will be a problem; currently, subscriptions act
as children of both queues _and_ messages. I may have to rework our
semantics around that.

I also need to decide how to handle deleting multiple messages at
once/clearing a queue. My gut says they should be the same endpoint,
just with an ID filter in the URL, but that will require more thought.