Updated resource files to new spec.
Updated our spec to:
* Clarify ownership of requirements. Required properties are now
defined with the properties themselves, and not part of interactions.
Interactions simply expect all required properties. This matches the
semantics of a property much more cleanly. Rather than specifying a
property as a requirement, any property without a default value is
considered required.
* Manage r/w permissions more simply, by just specifying whether a
 property is readable or writable in a permissions array.
* Make URL construction easier and more flexible by including the URL
 prefix as a property and specifying which property we want to use as
the slug.

Fleshed out all MQ endpoints using the new semantics, except for
deleting groups of messages all at once. Need to think carefully about
the semantics for that.

Updated tooling to read the entire resource tree (reading parents, etc.)
and produce a list of API endpoints on demand.