Spec out deleting many messages/clearing a queue.
Added an interaction for deleting multiple messages from MQ in a single
request. This interaction can also clear the queue, when no message IDs
are included.

This required adding the "accept_many" property to interactions, which,
when set to true, will tell the interaction to expect multiple resources
in the request. It will also prevent the url_slug from being included in
URL endpoints.

This also required adding the "repeated" property to properties, which,
when set to true, will tell interactions that the property may be seen
multiple times in a single request. This should only ever occur in URL
parameters.

The apidef-test.go file has been updated to display non-optional
parameters as part of the endpoints.

The apidef package was updated to take advantage of the new properties.