Add support for MQ's push queues.
Added sample resource documents for the push queue resources necessary.
I'm still not 100% sold on the naming, and have some concerns about
removing multiple subscribers from a queue in a single request (will the
URLs all need to be in the querystring? If so, ew.) but this brings MQ
up to feature-completion.