Start generating example requests.
The endpoint.SampleRequest slice of bytes is now populated by randomly
generated data that fits the constraints of the resource and
interaction.

This is not implemented for pointer types, and it should be.

This is not implemented for array types or object types, and I'm not
sure those should be types. Arrays of what? Objects of what? I think
array should just be signaled with the Repeated property, and objects
should just be new resources. Still thinking on that. So far, neither
type is needed in the MQ spec.

I still need to implement generating a random Time and a random float.
Right now, the zero values are used.