Don't provide a default ID when deleting messages.
I don't think the ID should be considered optional, I think that the
"accepts many" in the params should inherently mean that omission is
okay. This allows us to include them in the sample requests without
cluttering sample requests unduly.

Still not sure about this one, as it's conflating two separate ideas.
I'm still not comfortable with the semantics around deleting many
messages, really. I'd be interested to know how necessary this endpoint
really is. I'm not sure it should exist.