Updated resource files spec to include nil values.
We want to be able to specify that a value is, by default, not set, so
we need the special word "nil" to be defined in the spec. Previously,
the option was to just omit the default value, which would mark the
property as required, or to set the default value to the zero value of
its type, which is not the same semantically.